Split out input_metadata.cpp from input_metadata.h
[ghstack-poisoned]
